+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 
 ## [Unreleased]
 
+### Fixed
+
+- Clear the cached file tree when the project changes and the project git repository on the server is synced.
+
 
 
 ## [1.2.0] – 2025-03-27